+++ app/src/main/java/com/example/pengingatku/utils/ScreenNavigation.kt
@@ -38,7 +38,7 @@ sealed class ScreenNavigation(val route: String, val title: String, val icon: In
     // Other Screens
     object AddTimer : ScreenNavigation("add_timer", "Add Timer")
     object EditTimer : ScreenNavigation("edit_timer/{timerId}", "Edit Timer") {
-        fun createRoute(id: Int) = "edit_timer/$id"
+        fun createRoute(id: Long) = "edit_timer/$id"
     }
 
     companion object {
